+++ src/api/player.py
@@ -28,6 +28,17 @@ class Player:
 			self.wrapper.server.console(string)
 		except:
 			pass
+	def execute(self, string):
+		""" Run a vanilla command as this player. Works best in proxy mode. If proxy mode is not enabled, it simply falls back to using the 1.8 'execute' command. 
+		
+		To be clear, this does NOT work with any Wrapper.py commands. The command is sent straight to the vanilla server."""
+		try:
+			self.client.message("/%s" % string)
+		except:
+			self.console("execute %s ~ ~ ~ %s" % string)
+	def say(self, string):
+		""" Send a message as a player. Beware, as this does not filter commands, so it could be used to execute commands as the player. Only works in proxy mode. """
+		self.client.message(string)
 	def getClient(self):
 		if self.client == None:
 			for client in self.wrapper.proxy.clients: