+++ src/globals.py
@@ -1,2 +1,2 @@
-build=70
+build=71
 type='dev'