+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 #Changelog#
-
+Fix issue with [23:10:26] [Wrapper.py/ERROR] Traceback (most recent call last):
+[23:10:26] [Wrapper.py/ERROR]   File "Wrapper.py/__main__.py", line 136, in callEvent
+[23:10:26] [Wrapper.py/ERROR]     result = self.events[pluginID][event](payload)
+[23:10:26] [Wrapper.py/ERROR]   File "Wrapper.py/web.py", line 45, in onPlayerMessage
+[23:10:26] [Wrapper.py/ERROR]     del self.consoleScrollback[0]
 <h4>0.7.5</h4>
 **Features**
 - Web interface improvements:
@@ -19,6 +23,8 @@
 
 **Developer Changes**
 - "AUTHOR" and "WEBSITE" plugin metadata variables added (see template.py for example)
+- player.say(message): Say something through the player. Proxy mode only.
+- player.execute(command): Execute a command as the player. Works best in proxy mode, but will fallback to using the 1.8 'execute' command if proxy mode is not available.
 
 <h4>0.7.4</h4>
 Just a small little update, to fix a few things, and improve upon some existing features.